JWT Authentication and Features Update

+++ src/main/java/com/contractmanagementsystem/service/ClientService.java
@@ -2,26 +2,27 @@ package com.contractmanagementsystem.service;
 
 import com.contractmanagementsystem.dto.ContractRequestDTO;
 import com.contractmanagementsystem.dto.ContractResponseDTO;
+import com.contractmanagementsystem.dto.QuestionAnswerResponseDTO;
 import com.contractmanagementsystem.exception.ContractException;
+import com.contractmanagementsystem.exception.UserException;
 import com.contractmanagementsystem.model.Contract;
 import com.contractmanagementsystem.model.ContractStatus;
 import com.contractmanagementsystem.repository.ContractRepository;
+import com.contractmanagementsystem.utils.AskQuestionUtil;
+import com.contractmanagementsystem.utils.DownloadFileUtil;
 import com.contractmanagementsystem.utils.FileStorageUtil;
 import com.contractmanagementsystem.utils.TextExtractionUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import org.springframework.core.io.Resource;
-import org.springframework.core.io.UrlResource;
-import org.springframework.http.HttpHeaders;
 
 import java.net.MalformedURLException;
+import java.util.Objects;
 
 
 @RequiredArgsConstructor
@@ -31,10 +32,12 @@ public class ClientService {
 
     private final TextExtractionUtil textExtractionUtil;
     private final FileStorageUtil fileStorageUtil;
+    private final DownloadFileUtil downloadFileUtil;
+    private final AskQuestionUtil askQuestionUtil;
 
 
 
-    public ResponseEntity<Map<String, Object>> uploadContract(ContractRequestDTO dto) throws Exception {
+    public ResponseEntity<Map<String, Object>> uploadContract(ContractRequestDTO dto, String id, String clientName) throws Exception {
 
         Map<String, Object> response = new HashMap<>();
 
@@ -49,6 +52,12 @@ public class ClientService {
         contract.setContractPath(filePath);
 
         contract.setExtractedText(extractedText);
+
+        contract.setClientId(id);
+
+        contract.setClientName(clientName);
+
+        contract.setConsultantId(dto.getConsultantId());
 
         contract.setStatus(ContractStatus.DRAFT);
 
@@ -67,11 +76,15 @@ public class ClientService {
                 .body(response);
     }
 
-    public ResponseEntity<Map<String, Object>> updateContract(String id, ContractRequestDTO dto) throws Exception {
+    public ResponseEntity<Map<String, Object>> updateContract(String userId,String id, ContractRequestDTO dto) throws Exception {
 
         Map<String, Object> response = new HashMap<>();
 
         Contract contract = contractRepository.findById(id).orElseThrow(() -> new ContractException("No Contract with this id exists"));
+
+        if(!Objects.equals(contract.getClientId(), userId)){
+            throw new UserException("This Contract don't belong to you So you can't update it");
+        }
 
         // Delete old file
         fileStorageUtil.deleteFile(contract.getContractPath());
@@ -86,6 +99,8 @@ public class ClientService {
         contract.setContractPath(filePath);
 
         contract.setExtractedText(extractedText);
+
+        contract.setConsultantId(dto.getConsultantId());
 
         contract.setStatus(ContractStatus.DRAFT);
 
@@ -102,34 +117,63 @@ public class ClientService {
         return ResponseEntity.ok(response);
     }
 
-    public ResponseEntity<Resource> downloadContract(String id) throws MalformedURLException {
+    public ResponseEntity<Map<String, Object>>getAllContracts(String id) {
 
-        Contract contract = contractRepository.findById(id).orElseThrow(() -> new ContractException("Contract not found"));
+        List<Contract> contracts = contractRepository.findByClientId(id);
 
-        Path path = Paths.get(contract.getContractPath());
+        Map<String, Object> response = new HashMap<>();
 
-        Resource resource = new UrlResource(path.toUri());
+        response.put("message",contracts.isEmpty() ? "No contracts found" : "Contracts fetched successfully");
+        List<ContractResponseDTO> responseDTOList=contracts.stream()
+                .map(contract -> {
+                    ContractResponseDTO dto = new ContractResponseDTO();
 
-        if (!resource.exists()) {
-            throw new ContractException("File not found");
-        }
+                    dto.setId(contract.getId());
 
-        String fileName = resource.getFilename();
-        MediaType mediaType;
+                    dto.setContractName(contract.getContractName());
 
-        if (fileName != null && fileName.endsWith(".pdf")) {
-            mediaType = MediaType.APPLICATION_PDF;
-        }
-        else if (fileName != null && fileName.endsWith(".docx")) {
-            mediaType = MediaType.parseMediaType("application/vnd.openxmlformats-officedocument.wordprocessingml.document");
-        }
-        else {
-            throw new ContractException("Only PDF and DOCX files are allowed");
-        }
+                    dto.setStatus(contract.getStatus());
+                    return dto;
+                })
+                .toList();
 
-        return ResponseEntity.ok().contentType(mediaType).
-                header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + fileName + "\"").
-                body(resource);
+        response.put("data",responseDTOList);
+
+        return ResponseEntity.ok(response);
     }
+
+    public ResponseEntity<Resource> downloadFile(String clientId, String id) throws MalformedURLException {
+        Contract contract=contractRepository.findById(id).orElseThrow(()->new ContractException("No Such Contract Exists"));
+        if(!Objects.equals(contract.getClientId(),clientId)){
+            throw new UserException("You are not Authorized to access this");
+        }
+        return downloadFileUtil.downloadContract(contract);
+
+    }
+
+    public ResponseEntity<Map<String,Object>>getContract(String clientId,String id){
+        Contract contract=contractRepository.findById(id).orElseThrow(() -> new ContractException("No Such Contract Exist"));
+        if(!Objects.equals(contract.getClientId(),clientId)){
+            throw new UserException("You are not Authorized to access this");
+        }
+        ContractResponseDTO responseDTO=new ContractResponseDTO();
+        responseDTO.setId(contract.getId());
+        responseDTO.setContractName(contract.getContractName());
+        responseDTO.setStatus(contract.getStatus());
+        Map<String,Object> response=new HashMap<>();
+        response.put("message","Contract details Successfully fetched");
+        response.put("Contract Details",responseDTO);
+        return ResponseEntity.ok().body(response);
+    }
+
+    public QuestionAnswerResponseDTO askQuestion(String clientId, String id, String question){
+        Contract contract = contractRepository.findById(id).orElseThrow(() -> new ContractException("No Contract With This ID exist"));
+        if (!Objects.equals(contract.getClientId(), clientId)) {
+            throw new UserException("You are not Authorized to perform this operation");
+        }
+        return askQuestionUtil.askQuestion(contract,question);
+
+    }
+
 }
